field-not-required

+++ lib/backend/sanitize.go
@@ -20,8 +20,9 @@ package backend
 
 import (
 	"context"
-	"iter"
+	"regexp"
 	"time"
+	"unicode/utf8"
 
 	"github.com/gravitational/trace"
 	"github.com/jonboulle/clockwork"
@@ -30,59 +31,36 @@ import (
 // errorMessage is the error message to return when invalid input is provided by the caller.
 const errorMessage = "special characters are not allowed in resource names, please use name composed only from characters, hyphens, dots, and plus signs: %q"
 
-// isValidKeyByte checks if the byte is a valid character for a key.
-func isValidKeyByte(b byte) bool {
-	switch b {
-	case
-		// Digits
-		'0', '1', '2', '3', '4', '5', '6', '7', '8', '9',
+// allowPattern is the pattern of allowed characters for each key within
+// the path.
+var allowPattern = regexp.MustCompile(`^[0-9A-Za-z@_:.\-/+]*$`)
 
-		// Lowercase letters
-		'a', 'b', 'c', 'd', 'e', 'f', 'g', 'h', 'i', 'j', 'k', 'l', 'm',
-		'n', 'o', 'p', 'q', 'r', 's', 't', 'u', 'v', 'w', 'x', 'y', 'z',
-
-		// Uppercase letters
-		'A', 'B', 'C', 'D', 'E', 'F', 'G', 'H', 'I', 'J', 'K', 'L', 'M',
-		'N', 'O', 'P', 'Q', 'R', 'S', 'T', 'U', 'V', 'W', 'X', 'Y', 'Z',
-
-		// Allowed symbols
-		'@', '_', ':', '.', '-', '+':
-		return true
-	default:
-		return false
-	}
+// denyPattern matches some unallowed combinations
+var denyPatterns = []*regexp.Regexp{
+	regexp.MustCompile(`//`),
+	regexp.MustCompile(`(^|/)\.\.?(/|$)`),
 }
 
 // IsKeySafe checks if the passed in key conforms to whitelist
-func IsKeySafe(key Key) bool {
-	components := key.Components()
-	for i, k := range components {
-		switch k {
-		case noEnd:
-			continue
-		case ".", "..":
-			return false
-		case "":
-			return key.exactKey && i == len(components)-1
-		}
+func IsKeySafe(s Key) bool {
+	return allowPattern.Match(s) && !denyPatternsMatch(s) && utf8.Valid(s)
+}
 
-		for _, b := range []byte(k) {
-			if !isValidKeyByte(b) {
-				return false
-			}
+// denyPatternsMatch checks if the passed in key conforms to the deny patterns.
+func denyPatternsMatch(s Key) bool {
+	for _, pattern := range denyPatterns {
+		if pattern.Match(s) {
+			return true
 		}
 	}
-	return true
+
+	return false
 }
 
 var _ Backend = (*Sanitizer)(nil)
 
-// Sanitizer wraps a [Backend] implementation to make sure all
-// [Key]s written to the backend are allowed. Retrieval and deletion
-// of items do not perform any [Key] sanitization in order to allow
-// interacting with any items that might already exist in the
-// [Backend] prior to validation being performed on each
-// subcomponent of a [Key] instead of on the entire [Key].
+// Sanitizer wraps a Backend implementation to make sure all values requested
+// of the backend are whitelisted.
 type Sanitizer struct {
 	backend Backend
 }
@@ -96,11 +74,10 @@ func NewSanitizer(backend Backend) *Sanitizer {
 
 // GetRange returns query range
 func (s *Sanitizer) GetRange(ctx context.Context, startKey, endKey Key, limit int) (*GetResult, error) {
+	if !IsKeySafe(startKey) {
+		return nil, trace.BadParameter(errorMessage, startKey)
+	}
 	return s.backend.GetRange(ctx, startKey, endKey, limit)
-}
-
-func (s *Sanitizer) Items(ctx context.Context, params ItemsParams) iter.Seq2[Item, error] {
-	return s.backend.Items(ctx, params)
 }
 
 // Create creates item if it does not exist
@@ -142,6 +119,9 @@ func (s *Sanitizer) ConditionalUpdate(ctx context.Context, i Item) (*Lease, erro
 
 // Get returns a single item or not found error
 func (s *Sanitizer) Get(ctx context.Context, key Key) (*Item, error) {
+	if !IsKeySafe(key) {
+		return nil, trace.BadParameter(errorMessage, key)
+	}
 	return s.backend.Get(ctx, key)
 }
 
@@ -157,16 +137,28 @@ func (s *Sanitizer) CompareAndSwap(ctx context.Context, expected Item, replaceWi
 
 // Delete deletes item by key
 func (s *Sanitizer) Delete(ctx context.Context, key Key) error {
+	if !IsKeySafe(key) {
+		return trace.BadParameter(errorMessage, key)
+	}
 	return s.backend.Delete(ctx, key)
 }
 
 // ConditionalDelete deletes the item by key if the revision matches the stored revision.
 func (s *Sanitizer) ConditionalDelete(ctx context.Context, key Key, revision string) error {
+	if !IsKeySafe(key) {
+		return trace.BadParameter(errorMessage, key)
+	}
 	return s.backend.ConditionalDelete(ctx, key, revision)
 }
 
 // DeleteRange deletes range of items
 func (s *Sanitizer) DeleteRange(ctx context.Context, startKey, endKey Key) error {
+	// we only validate the start key, since we often compute the end key
+	// in order to delete a bunch of related entries
+	if !IsKeySafe(startKey) {
+		return trace.BadParameter(errorMessage, startKey)
+	}
+
 	return s.backend.DeleteRange(ctx, startKey, endKey)
 }
